Moved object ordering resposibility directly to Drawable
