+++ tests/test_element.py
@@ -1,5 +1,5 @@
 from .base_class import VisuscriptTestCase
-from visuscript.drawable.mixins import Element
+from visuscript.primatives.mixins import Element
 from visuscript.primatives import Vec2, Vec3
 from visuscript.config import config
