+++ visuscript/config.py
@@ -1,7 +1,7 @@
 import sys
 
 from visuscript.constants import OutputFormat
-from visuscript.primatives import Color
+from visuscript.primatives.mixins.color import Color
 
 
 class _AnimationConfig:
