+++ visuscript/primatives/mixins/__init__.py
@@ -1,2 +1,17 @@
-from .mixins import *
-from .color import *
+from .color import (
+    RgbMixin,
+    OpacityMixin,
+    Color)
+from .mixins import (
+    Drawable,
+    TransformMixin,
+    FillMixin,
+    StrokeMixin,
+    ShapeMixin,
+    TransformableShapeMixin,
+    AnchorMixin,
+    HierarchicalDrawable,
+    GlobalShapeMixin,
+    Element,
+    Shape,
+)
